♻️ Extract UI configuration to separate files
