+++ modules/ppcp-settings/resources/js/data/configuration.js
@@ -1,0 +1,10 @@
+export { BUSINESS_TYPES, PRODUCT_TYPES } from './onboarding/configuration';
+
+export {
+	STYLING_LOCATIONS,
+	STYLING_PAYMENT_METHODS,
+	STYLING_LABELS,
+	STYLING_COLORS,
+	STYLING_LAYOUTS,
+	STYLING_SHAPES,
+} from './styling/configuration';
