--- modules/ppcp-settings/resources/js/data/constants.js
@@ -1,9 +1,0 @@
-export { BUSINESS_TYPES, PRODUCT_TYPES } from './onboarding/constants';
-
-export {
-	STYLING_LOCATIONS,
-	STYLING_LABELS,
-	STYLING_COLORS,
-	STYLING_LAYOUTS,
-	STYLING_SHAPES,
-} from './styling/constants';
